LPD-57471 Create component to handle reusable space members management
