+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/css/spaces/SpaceMembersWithList.scss
@@ -1,4 +1,4 @@
-.add-space-members {
+.space-members-with-list {
 	padding: 24px;
 
 	.members-list {
